feat(code): huge refactor

+++ judge/worker/enviro.py
@@ -3,8 +3,9 @@ import os
 import shutil
 from tempfile import mkdtemp
 
-from judge.utils import is_debug
 from judge.log import get_logger
+from judge.runner import CaseConfig
+from judge.utils import is_debug
 
 
 class Environment(object):
@@ -17,6 +18,7 @@ class Environment(object):
         self.task = task
         self._current_dir = os.getcwd()
         self._prepare_working_dir()
+        self._write_code()
         self.write_case_config()
 
     def _prepare_working_dir(self):
@@ -26,23 +28,30 @@ class Environment(object):
 
         os.chdir(self.path)
 
-    def place_user_input(self, data):
+    def _write_code(self):
+        with open(self.task.language_type.source_name, 'w') as f:
+            f.write(self.task.code)
+
+    def place_input(self, data):
         # type: (str) -> None
-        inf = open('user.in', 'w+')
-        inf.write(data)
-        inf.close()
+        with open('user.in', 'w+') as f:
+            f.write(data)
+
+    def write_compile_config(self):
+        with open('compile.json', 'w+') as f:
+            content = json.dumps(self.task.language_type.to_compile_info())
+            f.write(content)
 
     def write_case_config(self):
-        file = open("case.json", "w")
-        content = json.dumps(self.task.as_task_info())
-        file.write(content)
-        file.close()
+        config = CaseConfig(self.task)
+        config.write_to_file()
 
     def prepare_for_next(self):
         get_logger().info("Clear working dir for next case")
         files = ['user.in', 'user.out', 'user.err']
         for file in files:
-            os.unlink(file)
+            if os.path.exists(file):
+                os.unlink(file)
 
     def clean(self):
         if self.path and not is_debug():
